+++ Bencodex/Types/IValue.cs
@@ -16,6 +16,10 @@ namespace Bencodex.Types
     /// <seealso cref="Dictionary"/>
     public interface IValue : IEquatable<IValue>
     {
+        /// <summary>The number of bytes used for serializing the value.</summary>
+        [Pure]
+        int EncodingLength { get; }
+
         /// <summary>A JSON-like human-readable representation for
         /// debugging.</summary>
         /// <returns>A JSON-like representation.</returns>